[android] Moved purchase callback classes from bookmark payment fragment
